channelz: use grpc web for angular app (#296)
The grpc-web generated library is loaded in the html page via a
<script src="...">, and from the angular app's point of view, the
proto variable is just an arbitrary external javascript handle.

The envoy proxy host+port is controlled by environment.ts
and environment.prod.ts. This may change if it proves problematic
when integrating with envoy.

The binary protos must be manually turned into human friendly strings,
because js protos can not be dumped as JSONs.